+++ lib/squadron_service.dart
@@ -10,4 +10,4 @@ export 'src/squadron_exception.dart' show SquadronException;
 export 'src/worker_exception.dart'
     show WorkerException, CancelledException, TaskTimeoutException;
 export 'src/worker_service.dart' show CommandHandler, WorkerService;
-export 'src/squadron_serializer.dart';
+export 'src/squadron_marshaller.dart';