+++ lib/src/annotations/serialize_with.dart
@@ -1,13 +1,14 @@
-import '../squadron_serializer.dart';
+import '../squadron_marshaller.dart';
 
 /// Annotation used to indicate the type of a serializer to be used when
 /// marshalling/unmarshalling data to/from a worker service. Package
 /// `squadron_builder` will generate appropriate serialization code based
-/// on the provided [serializer].
+/// on the provided [marshaller].
 class SerializeWith {
-  const SerializeWith(this.serializer);
+  const SerializeWith(this.marshaller);
 
-  /// Type of the serializer to use when marshalling/unmarshalling data to/from
-  /// the worker service. This type must implement [SquadronSerializer].
-  final Type serializer;
+  /// Type or constant instance of the marshaller to use for data serialization
+  /// to/from the worker service. The type or instance type must implement
+  /// [SquadronMarshaller].
+  final dynamic marshaller;
 }